break things up a bit and handle most of the rubocops
only thing it doesn't like is Metrics/CyclomaticComplexity
and Metrics/AbcSize which is ugh. fuck you too rubocop
