+++ app/policies/bookmark_searcher_policy.rb
@@ -1,7 +1,7 @@
 class BookmarkSearcherPolicy < ApplicationPolicy
   class Scope < Scope
     def resolve
-      scope.new.query({ term: { user_id: user.id } })
+      scope.new.query(term: { user_id: user.id })
     end
   end
 end
