+++ app/searchers/application_searcher/query.rb
@@ -4,7 +4,7 @@ class ApplicationSearcher
 
     def initialize clauses
       @clauses = clauses
-      grouped = @clauses.chunk { |c| c.operator }.to_h
+      grouped = @clauses.chunk(&:operator).to_h
 
       @should_terms   = grouped.fetch(:should, [])
       @must_terms     = grouped.fetch(:must, [])
@@ -31,13 +31,14 @@ class ApplicationSearcher
         should_terms = fields.map do |field|
           clause.dup.tap do |dup_clause|
             dup_clause.field = field
+            dup_clause.op = nil
             dup_clause.operator = :should
-            dup_clause.op = '+'
           end
         end
 
         NestedClause.new clause.op, should_terms
-      end.chunk { |c| c.operator }.to_h
+      end
+        .chunk(&:operator).to_h
 
       @should_terms   = grouped_expanded_clauses.fetch(:should, [])
       @must_terms     = grouped_expanded_clauses.fetch(:must, [])
